More clang fixes: laptop mouse-poll, 64-bit ptr casts, narrowings
- Laptop/laptop.cpp: two more GetCursorPos / ScreenToClient sites
  gated behind _WIN32 with the gusMouseXPos / gusMouseYPos fallback.
- Strategic/Strategic Movement.cpp: three (UINT32)ptr casts widened
  to (UINT32)(uintptr_t)ptr. These are pre-existing 64-bit truncation
  bugs that MSVC silently accepted; the intermediate uintptr_t cast
  acknowledges the truncation so clang stops warning. Real fix is
  widening the dialogue-event API to uintptr_t but that's invasive.
- Tactical/Soldier Control.cpp: same (UINT32)(uintptr_t) treatment
  for one this-pointer cast, plus narrowing-cast fixes on the
  gsFullTileDirections INT16 initializer (the values are derived
  from runtime WORLD_COLS so the brace narrowing rule fires).
- TileEngine/Isometric Utils.cpp: same narrowing-cast pattern on the
  DirIncrementer initializer.

+++ Tactical/Soldier Control.cpp
@@ -1704,7 +1704,7 @@ BOOLEAN gfCalcTranslucency = FALSE;
 
 INT16		gsFullTileDirections[MAX_FULLTILE_DIRECTIONS] =
 {
-	-1, -WORLD_COLS - 1, -WORLD_COLS
+	(INT16)-1, (INT16)(-WORLD_COLS - 1), (INT16)-WORLD_COLS
 
 };
 
@@ -2446,7 +2446,7 @@ BOOLEAN SOLDIERTYPE::CreateSoldierCommon( UINT8 ubBodyType, SoldierID usSoldierI
 
 	if ( this->ubBodyType == QUEENMONSTER )
 	{
-		this->iPositionSndID = NewPositionSnd( NOWHERE, POSITION_SOUND_FROM_SOLDIER, (UINT32)this, QUEEN_AMBIENT_NOISE, 15 );
+		this->iPositionSndID = NewPositionSnd( NOWHERE, POSITION_SOUND_FROM_SOLDIER, (UINT32)(uintptr_t)this, QUEEN_AMBIENT_NOISE, 15 );
 	}
